feat(third-party) Add third party
Add third party: Lodash

We're simply going to run npm install like we would any other package.
Now one thing to keep in mind here is that we're building this
with TypeScript, so it'd be nice to have some typings files.
Some JavaScript libraries ship with the TypeScript definition files,
but not lodash, so we want to add typings files for that.
We do this by using the typings tool.